Made new Class with all Functions from Main
