+++ src/test/MainTest.java
@@ -3,6 +3,7 @@ package test;
 import main.Main;
 import main.CsvData;
 import main.SynonymMapBuilder;
+import main.UtilitiesFunction;
 import org.junit.jupiter.api.BeforeEach;
 
 import java.io.IOException;
@@ -29,13 +30,13 @@ class MainTest {
         CsvData testData = new CsvData();
         testData.init();
         List<Map<String,String>> accoms = testData.buildObject("src/main/info.csv");
-        String filterMap = Main.filterAccoms(accoms, FILTERS).toString();
+        String filterMap = UtilitiesFunction.filterAccoms(accoms, FILTERS).toString();
         assert(filterMap.contains("Highfield Park"));
         assert(!filterMap.contains("Dominick Place"));
 
         FILTERS.clear();
         FILTERS.put("Amenities", "disable gymnasium television room");
-        filterMap = Main.filterAccoms(accoms, FILTERS).toString();
+        filterMap = UtilitiesFunction.filterAccoms(accoms, FILTERS).toString();
         assert(filterMap.contains("Dominick Place") && filterMap.contains("Beckett House"));
         //System.out.println(FILTERS.keySet().toString());
         assert(FILTERS.containsKey("Disability_Access"));
